requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
certifi==2022.5.18.1
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.9
django-cors-headers==3.12.0
django-crontab==0.7.1
django-filter==2.3.0
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.1
getmac==0.8.2
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itypes==1.2.0
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyfcm==1.5.4
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-http-client==3.3.7
pytz==2022.1
redis==4.3.1
requests==2.27.1
rsa==4.7.2
sendgrid==6.9.5
sentry-sdk==1.4.3
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.1
zope.interface==5.4.0
